Put inital setup code in application.py instead of init file
